Create README.md
"4-bit Processor" is a Digital Logic Design (DLD) project that demonstrates the design and implementation of a basic processor architecture. This project showcases the application of digital logic principles in building a functional processor capable of executing simple instructions. It highlights my proficiency in designing digital systems and understanding processor architecture.